+++ src/apiClient.ts
@@ -107,7 +107,9 @@ export default class APIClient {
           }
           url.searchParams.set('metadata_pair', metadataPair.join(','));
         } else if (Array.isArray(value)) {
-          url.searchParams.append(snakeCaseKey, value.join(','));
+          for (const item of value) {
+            url.searchParams.append(snakeCaseKey, item as string);
+          }
         } else if (typeof value === 'object') {
           for (const item in value) {
             url.searchParams.append(