+++ tests/apiClient.spec.ts
@@ -115,7 +115,23 @@ describe('APIClient', () => {
 
         expect(options.url).toEqual(
           new URL(
-            'https://api.us.nylas.com/test?foo=bar&list=a%2Cb%2Cc&map=key1%3Avalue1&map=key2%3Avalue2'
+            'https://api.us.nylas.com/test?foo=bar&list=a&list=b&list=c&map=key1%3Avalue1&map=key2%3Avalue2'
+          )
+        );
+      });
+
+      it('should handle repeated query parameters', () => {
+        const options = client.requestOptions({
+          path: '/test',
+          method: 'GET',
+          queryParams: {
+            eventType: ['default', 'outOfOffice', 'focusTime'],
+          },
+        });
+
+        expect(options.url).toEqual(
+          new URL(
+            'https://api.us.nylas.com/test?event_type=default&event_type=outOfOffice&event_type=focusTime'
           )
         );
       });